Add attr_readers and initialize to controller, and make name better suited for use
